+++ app/src/main/kotlin/org/strigate/ferrot/app/receiver/BootCompletedReceiver.kt
@@ -43,7 +43,7 @@ class BootCompletedReceiver : BroadcastReceiver() {
                     .getAutomaticDuplicateDownloadDeletionSettingAsFlowUseCase()
                     .first()
                 if (automaticDuplicateDownloadDeletionSetting) {
-                    DeleteAllDuplicateDownloadsWorker.enqueueDebouncedCleanup(context)
+                    DeleteAllDuplicateDownloadsWorker.enqueueDebouncedReplace(context)
                 }
             }
         }