+++ app/src/main/kotlin/org/strigate/ferrot/domain/usecase/DownloadAudioUseCase.kt
@@ -1,6 +1,7 @@
 package org.strigate.ferrot.domain.usecase
 
 import org.strigate.ferrot.domain.usecase.downloadaudio.DeleteDownloadAudioUseCase
+import org.strigate.ferrot.domain.usecase.downloadaudio.GetAllDownloadAudioFilePathsUseCase
 import org.strigate.ferrot.domain.usecase.downloadaudio.GetDownloadAudioByDownloadIdAsFlowUseCase
 import org.strigate.ferrot.domain.usecase.downloadaudio.SaveDownloadAudioUseCase
 import javax.inject.Inject
@@ -9,4 +10,5 @@ class DownloadAudioUseCase @Inject constructor(
     val saveDownloadAudioUseCase: SaveDownloadAudioUseCase,
     val getDownloadAudioByDownloadIdAsFlowUseCase: GetDownloadAudioByDownloadIdAsFlowUseCase,
     val deleteDownloadAudioUseCase: DeleteDownloadAudioUseCase,
+    val getAllDownloadAudioFilePathsUseCase: GetAllDownloadAudioFilePathsUseCase,
 )