+++ app/src/main/kotlin/org/strigate/ferrot/domain/usecase/DownloadVideoUseCase.kt
@@ -1,6 +1,7 @@
 package org.strigate.ferrot.domain.usecase
 
 import org.strigate.ferrot.domain.usecase.downloadvideo.DeleteDownloadVideoUseCase
+import org.strigate.ferrot.domain.usecase.downloadvideo.GetAllDownloadVideoFilePathsUseCase
 import org.strigate.ferrot.domain.usecase.downloadvideo.GetDownloadIdsBySha256UseCase
 import org.strigate.ferrot.domain.usecase.downloadvideo.GetDownloadVideoByDownloadIdAsFlowUseCase
 import org.strigate.ferrot.domain.usecase.downloadvideo.SaveDownloadVideoUseCase
@@ -11,4 +12,5 @@ class DownloadVideoUseCase @Inject constructor(
     val getDownloadVideoByDownloadIdAsFlowUseCase: GetDownloadVideoByDownloadIdAsFlowUseCase,
     val getDownloadIdsBySha256UseCase: GetDownloadIdsBySha256UseCase,
     val deleteDownloadVideoUseCase: DeleteDownloadVideoUseCase,
+    val getAllDownloadVideoFilePathsUseCase: GetAllDownloadVideoFilePathsUseCase,
 )